# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive Driver Tracking SaaS Application for Indonesian Fleet Management",
        "title": "FleetTracker Pro API",
        "termsOfService": "https://fleettracker.id/terms",
        "contact": {
            "name": "FleetTracker Pro Support",
            "url": "https://fleettracker.id/support",
            "email": "support@fleettracker.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/analytics/fleet/maintenance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get maintenance scheduling insights and alerts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get maintenance insights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analytics.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/analytics.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/drivers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List drivers with filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "List drivers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driver status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Employment status",
                        "name": "employment_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Performance grade (A, B, C, D, F)",
                        "name": "performance_grade",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "City",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Province",
                        "name": "province",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Has vehicle assigned",
                        "name": "has_vehicle",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Is available",
                        "name": "is_available",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Is compliant",
                        "name": "is_compliant",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/driver.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new driver with Indonesian compliance validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Create a new driver",
                "parameters": [
                    {
                        "description": "Driver data",
                        "name": "driver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/driver.CreateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/driver.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/driver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "analytics.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "analytics.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "driver.CreateDriverRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "date_of_birth",
                "email",
                "first_name",
                "hire_date",
                "last_name",
                "nik",
                "phone_number",
                "province",
                "sim_number"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 10
                },
                "city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "hire_date": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "medical_checkup_date": {
                    "type": "string"
                },
                "nik": {
                    "description": "Indonesian Compliance Fields",
                    "type": "string"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "province": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "sim_expiry": {
                    "type": "string"
                },
                "sim_number": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 10
                },
                "training_completed": {
                    "type": "boolean"
                },
                "training_expiry": {
                    "type": "string"
                }
            }
        },
        "driver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "driver.Meta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_previous": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "driver.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/driver.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "driver.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}